requirements.txt
asgiref==3.6.0
Django==4.1.6
# django-rest==0.8.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
# six==1.16.0
sqlparse==0.4.3
whitenoise==6.3.0
